+++ src/FSharp.Data.DesignTime/Json/JsonProvider.fs
@@ -11,6 +11,7 @@ open FSharp.Data.Runtime
 open FSharp.Data.Runtime.BaseTypes
 open FSharp.Data.Runtime.StructuralTypes
 open FSharp.Data.Runtime.StructuralInference
+open System.Net
 
 // ----------------------------------------------------------------------------------------------
 
@@ -33,6 +34,11 @@ type public JsonProvider(cfg: TypeProviderConfig) as this =
         ProvidedTypeDefinition(asm, ns, "JsonProvider", None, hideObjectMethods = true, nonNullable = true)
 
     let buildTypes (typeName: string) (args: obj[]) =
+
+        // Enable TLS 1.2 for samples requested through https.
+        ServicePointManager.SecurityProtocol <-
+            ServicePointManager.SecurityProtocol
+            ||| SecurityProtocolType.Tls12
 
         // Generate the required type
         let tpType =