+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 * Change the namespace of the `InferenceMode` enum from `FSharp.Data.Runtime.StructuralInference` to `FSharp.Data`.
 * Move common runtime utilities out of `FSharp.Data.Http` and into a new `FSharp.Data.Runtime.Utilities` assembly.
 * Add `aria-label` to the list of html attributes used to infer names of types provided by the HtmlProvider.
+* Enable TLS 1.2 when requesting http(s) samples from the type providers.
 
 ### 6.0.1-beta001 - Aug 18 2022